requirements.txt
fastapi
uvicorn
python-dotenv
lxml
pandas
pandas-gbq
stomp.py
requests